+++ merged-packages/utils/src/node.test.ts
@@ -52,6 +52,7 @@ describe('node', () => {
         "VersionRangeStruct",
         "VersionStruct",
         "add0x",
+        "areUint8ArraysEqual",
         "assert",
         "assertExhaustive",
         "assertIsBytes",